// src/app/shared/models/models.model.ts
export class Ft_image {
    id?: string;
    url: string;
    ips?: string[];
    likes?: number;
    poster: string;
    city?: boolean;
    state?: boolean;
    country?: boolean;
    yearOfBirth?: boolean;
    active?: boolean;
    essence: string;
    footprint: string;
    timestamp?: number;
    isShow?: boolean;
    nindex?: number;
    footPrinted?: boolean;    
}